vip io

+++ src/io/input.rs
@@ -14,16 +14,16 @@ pub struct Input {
     
     /// width of the visible grid
     #[arg(short = 'W', long, default_value_t = Self::DEFAULT_WIDTH)]
-    pub width:  i32,
+    pub width:  usize,
     
     /// height of the visible grid
     #[arg(short = 'H', long, default_value_t = Self::DEFAULT_HEIGHT)]
-    pub height: i32,
+    pub height: usize,
 }
 
 impl Input {
     pub const DEFAULT_MINE_CONCENTRATION: f64 = 0.2f64;
     pub const DEFAULT_SEED: u64 = 0xDEADBEEF;
-    pub const DEFAULT_WIDTH:  i32 = 16;
-    pub const DEFAULT_HEIGHT: i32 = 16;
+    pub const DEFAULT_WIDTH:  usize = 16;
+    pub const DEFAULT_HEIGHT: usize = 16;
 }
